Introduce ImposterWorkflow to properly support operator workflows and GUWT.
Fixes #23, although ended up going with a slightly different approach than what was described in that issue.
Instead of making an interface that gives the workflow a key, removed the hard-coded reliance on the workflow's
`KClass` in `WorkflowId` into a strongly-typed `WorkflowIdentifier` type that is public in the core moddule,
along with an extension property to get the identifier for a workflow. `WorkflowIdentifier` is aware of wrapper
workflows via the `ImposterWorkflow` interface (instead of `KeyedWorkflow`). `WorkflowIdentifier` is
effectively a linked-list of workflow classes (although the classes may also be represented as strings, to
support rehydration from snapshots).