Introduce TreeSnapshot.
This is a concrete type that represents a complete snapshot of the entire workflow
tree, vs the snapshot of a single workflow's state. This makes the testing and runtime
API more typesafe, since it's impossible to accidentally pass a single workflow's
snapshot when a tree is expected. It also makes the internal code more clean because
it's clear which type of snapshot is expected where.

Closes #31 and unblocks https://github.com/square/workflow/issues/1192.